Added a gulp command generate-mod
It updates all config files with the stuff on packagage.json
Easy to start your own mod just by edditing that .json file.

Also added some info on the package.json file to match the content
that was on the files before.

It allows to update the copyright year too.